+++ packages/flame/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## [next]
+## [1.0.0-releasecandidate.17]
  - Added `StandardEffectController` class
  - Refactored `Effect` class to use `EffectController`, added `Transform2DEffect` class
  - Clarified `TimerComponent` example